Add files via upload
Array checker wasn't catching some arrays that were being passed to it, in the browser the array was seen as an object of type array and was bypassing the if statement.  Changed this check to Array.isArrray(displayData) and got inside of that if statement.  

Also changed inside that if statement because even after being inside array items werent being added to the <ul> properly. 
Declared a variable and initialized it as an empty array and pushed items that were being mapped to that array.