Added ContractDataTable and ContractDataTableRow
New components for displaying data from a smart contract in a
dynamicly updating table.

Often in Ethereum contracts, contract data is managed with a combination
of objects stored in a mapping and keys stored in an array. Thus to view
a list of objects, one must first retrieve the keys and then iterate
over them.

The ContractDataTable component included here solves this
problem by using a ContractDataTableRow component to generate a row for
each key.