Implemented real-time updates for ContractData
This commit solves the problem about ContractData not updating its value
even though methodArgs have changed. This is important for cases when,
for example, ContractData is used to call the balanceOf method of an
ERC20 token contract and methodArgs is uses the value of an input
element as a prop.

componentDidUpdate() is the right choice for updating the component
anytime the methodArgs change throughout ContractData's lifecycle as it
is invoked in every cycle, there are only 2 caveats:
- Always compare the previous methodArgs against the current methodArgs
of the
cycle so that setState() is not called on every single cycle.
When such thing happens, React throws:
```
Uncaught RangeError: Maximum call stack size exceeded
```
- Before such comparison is made, we need to first make sure methodArgs
is not undefined in the first place... Otherwise, the if statement will
compare two stringified undefined values, which will always be exactly
equal and will always fire the nested logic.